Adding the wallpaper stage
