+++ kano_init_flow/stages/wallpaper/__init__.py
@@ -1,0 +1,6 @@
+#
+# Copyright (C) 2015 Kano Computing Ltd.
+# License: http://www.gnu.org/licenses/gpl-2.0.txt GNU GPL v2
+#
+
+from .main import Wallpaper
